Add generic option to write ASReview data to RIS (#1634)
* Add generic option to write ASReview data to RIS

* Happy lint
